feat(student): 학생 전시 상세 페이지 내 ArtworkSection API 연동 (Video는 추후 Youtube Embed로 추가) (#73)
