--- src/pages/Exhibition/Student/Detail/ArtworkSection/ArtworkImages/ArtworkImages.tsx
@@ -1,31 +1,0 @@
-import { JSX } from 'react/jsx-runtime';
-
-import * as S from './ArtworkImages.styled';
-
-interface ArtworkImagesProps {
-  detailImages: {
-    id: number;
-    url: string;
-    caption: string;
-  }[];
-  currentPage: number;
-}
-
-const ArtworkImages = ({
-  detailImages,
-  currentPage,
-}: ArtworkImagesProps): JSX.Element => {
-  return (
-    <S.ArtworkImageContainer>
-      {detailImages.map((image) => (
-        <S.DetailImage
-          key={image.id}
-          src={`/Graduation-Exhibition/${currentPage}/${image?.url}`}
-          alt={image.caption}
-        />
-      ))}
-    </S.ArtworkImageContainer>
-  );
-};
-
-export default ArtworkImages;
