+++ src/pages/Exhibition/Student/Detail/ArtworkSection/ArtworkImages/StudentArtworkImages.styled.ts
