+++ src/pages/Exhibition/Student/Detail/ArtworkSection/StudentArtworkSection.styled.ts
