+++ src/pages/Exhibition/Student/Detail/ArtworkSection/VideoPlayer/StudentVideoPlayer.styled.ts
